+++ examples/run_CE_SSDO.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # ------------------------------------------------------------------------------------------------------%
-# Created by "Thieu Nguyen" at 12:00, 30/03/2020                                                        %
+# Created by "Thieu Nguyen" at 01:57, 27/04/2020                                                        %
 #                                                                                                       %
 #       Email:      nguyenthieu2102@gmail.com                                                           %
 #       Homepage:   https://www.researchgate.net/profile/Thieu_Nguyen6                                  %
@@ -8,18 +8,32 @@
 #-------------------------------------------------------------------------------------------------------%
 
 from opfunu.cec_basic.cec2014_nobias import *
-from opfunu.type_based.multi_modal import Functions
+from mealpy.probabilistic_based.CEM import CEBaseLCBO, CEBaseSSDO, CEBaseLCBONew
 from mealpy.human_based.LCBO import BaseLCBO, LevyLCBO, ImprovedLCBO
 
-t1 = Functions()
-
 ## Setting parameters
-objective_func = F22   #t1._pathological__
-problem_size = 100
+objective_func = F25
+problem_size = 50
 domain_range = [-100, 100]
 log = True
-epoch = 1000
+epoch = 50
 pop_size = 50
+
+md1 = CEBaseSSDO(objective_func, problem_size, domain_range, log, epoch, pop_size)
+best_pos1, best_fit1, list_loss1 = md1._train__()
+print(best_fit1)
+print("========================================================")
+
+md2 = CEBaseLCBO(objective_func, problem_size, domain_range, log, epoch, pop_size)
+best_pos2, best_fit2, list_loss2 = md2._train__()
+print(best_fit2)
+print("========================================================")
+
+md2 = CEBaseLCBONew(objective_func, problem_size, domain_range, log, epoch, pop_size)
+best_pos2, best_fit2, list_loss2 = md2._train__()
+print(best_fit2)
+print("========================================================")
+
 
 md1 = BaseLCBO(objective_func, problem_size, domain_range, log, epoch, pop_size)
 best_pos1, best_fit1, list_loss1 = md1._train__()
@@ -31,6 +45,8 @@ best_pos2, best_fit2, list_loss2 = md2._train__()
 print(best_fit2)
 print("========================================================")
 
-md3 = ImprovedLCBO(objective_func, problem_size, domain_range, log, epoch, pop_size)
-best_pos3, best_fit3, list_loss3 = md3._train__()
-print(best_fit3)
+md2 = ImprovedLCBO(objective_func, problem_size, domain_range, log, epoch, pop_size)
+best_pos2, best_fit2, list_loss2 = md2._train__()
+print(best_fit2)
+print("========================================================")
+